+++ addon/plugins/variable-plugin/variables/index.ts
@@ -2,3 +2,4 @@ export * from './number';
 export * from './text';
 export * from './codelist';
 export * from './location';
+export * from './address';